Neural Network - Digit Insight Project/requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.22
matplotlib>=3.7
tqdm>=4.65
gradio>=4.0
pillow>=9.0
pyyaml>=6.0
scipy>=1.11
